# interface/sdk/go/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Memory System
  description: HTTP API for HMS
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: HMS HTTP API
  version: 0.6.1
servers:
- url: /
paths:
...
  /v1/default/banks/{bank_id}/memories/list:
    get:
      description: "List memory units with pagination and optional full-text search.\
        \ Supports filtering by type. Results are sorted by most recent first (mentioned_at\
        \ DESC, then created_at DESC)."
      operationId: list_memories
      parameters:
      - explode: false
        in: path
        name: bank_id
        required: true
        schema:
          title: Bank Id
          type: string
        style: simple
      - explode: true
        in: query
        name: type
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: q
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: consolidation_state
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
        style: form
      - explode: false
        in: header
        name: authorization
        required: false
        schema:
          nullable: true
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMemoryUnitsResponse'
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List memory units
      tags:
      - Memory
...
  /v1/default/banks/{bank_id}/reflect:
    post:
      description: |-
        Reflect and formulate an answer using bank identity, world facts, and opinions.

        This endpoint:
        1. Retrieves experience (conversations and events)
        2. Retrieves world facts relevant to the query
        3. Retrieves existing opinions (bank's perspectives)
        4. Uses LLM to formulate a contextual answer
        5. Returns plain text answer and the facts used
      operationId: reflect
      parameters:
      - explode: false
        in: path
        name: bank_id
        required: true
        schema:
          title: Bank Id
          type: string
        style: simple
      - explode: false
        in: header
        name: authorization
        required: false
        schema:
          nullable: true
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReflectRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReflectResponse'
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reflect and generate answer
      tags:
      - Memory
...
components:
  schemas:
...
    Budget:
      description: Budget levels for recall/reflect operations.
      enum:
      - low
      - mid
      - high
      title: Budget
      type: string
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
          url: url
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
          url: url
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
      title: HTTPValidationError
...
    ListMemoryUnitsResponse:
      description: Response model for list memory units endpoint.
      example:
        items:
        - context: Work conversation
          date: 2024-01-15T10:30:00Z
          entities: "Alice (PERSON), Google (ORGANIZATION)"
          id: 550e8400-e29b-41d4-a716-446655440000
          text: Alice works at Google on the AI team
          type: world
        limit: 100
        offset: 0
        total: 150
      properties:
        items:
          items:
            additionalProperties: {}
          type: array
        total:
          title: Total
          type: integer
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
      required:
      - items
      - limit
      - offset
      - total
      title: ListMemoryUnitsResponse
...
    ReflectBasedOn:
      description: "Evidence the response is based on: memories, mental models, and\
        \ directives."
      properties:
        memories:
          default: []
          description: Memory facts used to generate the response
          items:
            $ref: '#/components/schemas/ReflectFact'
          type: array
        mental_models:
          default: []
          description: Mental models used during reflection
          items:
            $ref: '#/components/schemas/ReflectMentalModel'
          type: array
        directives:
          default: []
          description: Directives applied during reflection
          items:
            $ref: '#/components/schemas/ReflectDirective'
          type: array
      title: ReflectBasedOn
    ReflectDirective:
      description: A directive applied during reflect.
      properties:
        id:
          description: Directive ID
          title: Id
          type: string
        name:
          description: Directive name
          title: Name
          type: string
        content:
          description: Directive content
          title: Content
          type: string
      required:
      - content
      - id
      - name
      title: ReflectDirective
    ReflectFact:
      description: A fact used in think response.
      example:
        context: healthcare discussion
        id: 123e4567-e89b-12d3-a456-426614174000
        occurred_end: 2024-01-15T10:30:00Z
        occurred_start: 2024-01-15T10:30:00Z
        text: AI is used in healthcare
        type: world
      properties:
        id:
          nullable: true
          type: string
        text:
          description: "Fact text. When type='observation', this contains markdown-formatted\
            \ consolidated knowledge"
          title: Text
          type: string
        type:
          nullable: true
          type: string
        context:
          nullable: true
          type: string
        occurred_start:
          nullable: true
          type: string
        occurred_end:
          nullable: true
          type: string
      required:
      - text
      title: ReflectFact
    ReflectIncludeOptions:
      description: Options for including additional data in reflect results.
      properties:
        facts:
          description: Options for including facts (based_on) in reflect results.
          properties: {}
          title: FactsIncludeOptions
          type: object
        tool_calls:
          $ref: '#/components/schemas/ToolCallsIncludeOptions'
      title: ReflectIncludeOptions
    ReflectLLMCall:
      description: An LLM call made during reflect agent execution.
      properties:
        scope:
          description: "Call scope: agent_1, agent_2, final, etc."
          title: Scope
          type: string
        duration_ms:
          description: Execution time in milliseconds
          title: Duration Ms
          type: integer
      required:
      - duration_ms
      - scope
      title: ReflectLLMCall
    ReflectMentalModel:
      description: A mental model used during reflect.
      properties:
        id:
          description: Mental model ID
          title: Id
          type: string
        text:
          description: Mental model content
          title: Text
          type: string
        context:
          nullable: true
          type: string
      required:
      - id
      - text
      title: ReflectMentalModel
    ReflectRequest:
      description: Request model for reflect endpoint.
      example:
        budget: low
        include:
          facts: {}
        max_tokens: 4096
        query: What do you think about artificial intelligence?
        response_schema:
          properties:
            summary:
              type: string
            key_points:
              items:
                type: string
              type: array
          required:
          - summary
          - key_points
          type: object
        tags:
        - user_a
        tags_match: any
      properties:
        query:
          title: Query
          type: string
        budget:
          $ref: '#/components/schemas/Budget'
        context:
          nullable: true
          type: string
        max_tokens:
          default: 4096
          description: Maximum tokens for the response
          title: Max Tokens
          type: integer
        include:
          $ref: '#/components/schemas/ReflectIncludeOptions'
        response_schema:
          additionalProperties: {}
          nullable: true
        tags:
          items:
            type: string
          nullable: true
          type: array
        tags_match:
          default: any
          description: "How to match tags: 'any' (OR, includes untagged), 'all' (AND,\
            \ includes untagged), 'any_strict' (OR, excludes untagged), 'all_strict'\
            \ (AND, excludes untagged)."
          enum:
          - any
          - all
          - any_strict
          - all_strict
          title: Tags Match
          type: string
        tag_groups:
          items:
            $ref: '#/components/schemas/MentalModelTrigger_Input_tag_groups_inner'
          nullable: true
          type: array
        fact_types:
          items:
            enum:
            - world
            - experience
            - observation
            type: string
          nullable: true
          type: array
        exclude_mental_models:
          default: false
          description: "If true, exclude all mental models from the reflect loop (skip\
            \ search_mental_models tool)."
          title: Exclude Mental Models
          type: boolean
        exclude_mental_model_ids:
          items:
            type: string
          nullable: true
          type: array
      required:
      - query
      title: ReflectRequest
    ReflectResponse:
      description: Response model for think endpoint.
      example:
        based_on:
          memories:
          - id: "123"
            text: AI is used in healthcare
            type: world
          - id: "456"
            text: I discussed AI applications last week
            type: experience
        structured_output:
          key_points:
          - Used in healthcare
          - Discussed recently
          summary: AI is transformative
        text: |-
          ## AI Overview

          Based on my understanding, AI is a **transformative technology**:

          - Used extensively in healthcare
          - Discussed in recent conversations
          - Continues to evolve rapidly
        trace:
          llm_calls:
          - duration_ms: 1200
            scope: agent_1
          observations:
          - id: obs-1
            name: AI Technology
            subtype: structural
            type: concept
          tool_calls:
          - duration_ms: 150
            input:
              query: AI
            tool: recall
        usage:
          input_tokens: 1500
          output_tokens: 500
          total_tokens: 2000
      properties:
        text:
          description: "The reflect response as well-formatted markdown (headers,\
            \ lists, bold/italic, code blocks, etc.)"
          title: Text
          type: string
        based_on:
          $ref: '#/components/schemas/ReflectBasedOn'
        structured_output:
          additionalProperties: {}
          nullable: true
        usage:
          $ref: '#/components/schemas/TokenUsage'
        trace:
          $ref: '#/components/schemas/ReflectTrace'
      required:
      - text
      title: ReflectResponse
    ReflectToolCall:
      description: A tool call made during reflect agent execution.
      properties:
        tool:
          description: "Tool name: lookup, recall, learn, expand"
          title: Tool
          type: string
        input:
          additionalProperties: {}
          description: Tool input parameters
          title: Input
        output:
          additionalProperties: {}
          nullable: true
        duration_ms:
          description: Execution time in milliseconds
          title: Duration Ms
          type: integer
        iteration:
          default: 0
          description: Iteration number (1-based) when this tool was called
          title: Iteration
          type: integer
      required:
      - duration_ms
      - input
      - tool
      title: ReflectToolCall
    ReflectTrace:
      description: Execution trace of LLM and tool calls during reflection.
      properties:
        tool_calls:
          default: []
          description: Tool calls made during reflection
          items:
            $ref: '#/components/schemas/ReflectToolCall'
          type: array
        llm_calls:
          default: []
          description: LLM calls made during reflection
          items:
            $ref: '#/components/schemas/ReflectLLMCall'
          type: array
      title: ReflectTrace
...
    TagGroupAnd-Input:
      description: "Compound AND group: all child filters must match."
      properties:
        and:
          items:
            $ref: '#/components/schemas/MentalModelTrigger_Input_tag_groups_inner'
          type: array
      required:
      - and
      title: TagGroupAnd
...
    TagGroupLeaf:
      description: "A leaf tag filter: matches memories by tag list and match mode."
      example:
        match: any_strict
        tags:
        - tags
        - tags
      properties:
        tags:
          items:
            type: string
          type: array
        match:
          default: any_strict
          enum:
          - any
          - all
          - any_strict
          - all_strict
          title: Match
          type: string
      required:
      - tags
      title: TagGroupLeaf
    TagGroupNot-Input:
      description: "Compound NOT group: child filter must NOT match."
      properties:
        not:
          $ref: '#/components/schemas/Not'
      required:
      - not
      title: TagGroupNot
...
    TagGroupOr-Input:
      description: "Compound OR group: at least one child filter must match."
      properties:
        or:
          items:
            $ref: '#/components/schemas/MentalModelTrigger_Input_tag_groups_inner'
          type: array
      required:
      - or
      title: TagGroupOr
...
    TokenUsage:
      description: |-
        Token usage metrics for LLM calls.

        Tracks input/output tokens for a single request to enable
        per-request cost tracking and monitoring.
      example:
        input_tokens: 1500
        output_tokens: 500
        total_tokens: 2000
      properties:
        input_tokens:
          default: 0
          description: Number of input/prompt tokens consumed
          title: Input Tokens
          type: integer
        output_tokens:
          default: 0
          description: Number of output/completion tokens generated
          title: Output Tokens
          type: integer
        total_tokens:
          default: 0
          description: Total tokens (input + output)
          title: Total Tokens
          type: integer
      title: TokenUsage
    ToolCallsIncludeOptions:
      description: Options for including tool calls in reflect results.
      properties:
        output:
          default: true
          description: Include tool outputs in the trace. Set to false to only include
            inputs (smaller payload).
          title: Output
          type: boolean
      title: ToolCallsIncludeOptions
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        input: ""
        ctx: "{}"
        type: type
        url: url
      properties:
        loc:
          items:
            $ref: '#/components/schemas/ValidationError_loc_inner'
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input: {}
        ctx:
          title: Context
          type: object
        url:
          title: URL
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    MentalModelTrigger_Input_tag_groups_inner:
      anyOf:
      - $ref: '#/components/schemas/TagGroupLeaf'
      - $ref: '#/components/schemas/TagGroupAnd-Input'
      - $ref: '#/components/schemas/TagGroupOr-Input'
      - $ref: '#/components/schemas/TagGroupNot-Input'
...
    Not:
      anyOf:
      - $ref: '#/components/schemas/TagGroupLeaf'
      - $ref: '#/components/schemas/TagGroupAnd-Input'
      - $ref: '#/components/schemas/TagGroupOr-Input'
      - $ref: '#/components/schemas/TagGroupNot-Input'
      title: Not
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer